+++ enzyme/tools/enzyme-tblgen/enzyme-tblgen.cpp
@@ -24,506 +24,10 @@
 #include "llvm/TableGen/TableGenBackend.h"
 
 #include "caching.h"
+#include "general.h"
 
 // clang-format off
 //
-
-using namespace llvm;
-
-enum ActionType { GenDerivatives, GenBlasDerivatives };
-
-static cl::opt<ActionType>
-    action(cl::desc("Action to perform:"),
-           cl::values(clEnumValN(GenBlasDerivatives, "gen-blas-derivatives",
-                                 "Generate BLAS derivatives")),
-           cl::values(clEnumValN(GenDerivatives, "gen-derivatives",
-                                 "Generate instruction derivative")));
-
-bool hasDiffeRet(Init *resultTree) {
-  if (DagInit *resultRoot = dyn_cast<DagInit>(resultTree)) {
-    auto opName = resultRoot->getOperator()->getAsString();
-    auto Def = cast<DefInit>(resultRoot->getOperator())->getDef();
-    if (opName == "DiffeRet" || Def->isSubClassOf("DiffeRet")) {
-      return true;
-    }
-    for (auto arg : resultRoot->getArgs()) {
-      if (hasDiffeRet(arg))
-        return true;
-    }
-    for (auto zp :
-         llvm::zip(resultRoot->getArgs(), resultRoot->getArgNames())) {
-      if (hasDiffeRet(std::get<0>(zp)))
-        return true;
-    }
-  }
-  return false;
-}
-
-void getFunction(raw_ostream &os, std::string callval, std::string FT,
-                 std::string cconv, Init *func) {
-  if (DagInit *resultRoot = dyn_cast<DagInit>(func)) {
-    auto opName = resultRoot->getOperator()->getAsString();
-    auto Def = cast<DefInit>(resultRoot->getOperator())->getDef();
-    if (opName == "SameFunc" || Def->isSubClassOf("SameFunc")) {
-      os << "#if LLVM_VERSION_MAJOR >= 11\n";
-      os << "  auto " << callval << " = orig->getCalledOperand();\n";
-      os << "#else\n";
-      os << "  auto " << callval << " = orig->getCalledValue();\n";
-      os << "#endif\n";
-      os << "  auto " << FT << " = orig->getFunctionType();\n";
-      os << "  auto " << cconv << " = orig->getCallingConv();\n";
-      return;
-    }
-    if (opName == "SameTypesFunc" || Def->isSubClassOf("SameTypesFunc")) {
-      os << " auto " << FT << " = orig->getFunctionType();\n";
-      os << " auto " << callval
-         << " = gutils->oldFunc->getParent()->getOrInsertFunction(";
-      os << Def->getValueInit("name")->getAsString();
-      os << ", " << FT << ", called->getAttributes())\n";
-      os << "#if LLVM_VERSION_MAJOR >= 9\n";
-      os << "  .getCallee()\n";
-      os << "#endif\n";
-      os << ";\n";
-      os << "  auto " << cconv << " = orig->getCallingConv();\n";
-      return;
-    }
-  }
-  assert(0 && "Unhandled function");
-}
-void getIntrinsic(raw_ostream &os, std::string callval, std::string FT,
-                  std::string cconv, StringRef intrName, ListInit *typeInit,
-                  StringMap<std::string> &nameToOrdinal) {
-  os << " Type *tys[] = {";
-  bool first = true;
-  for (auto intrType : *typeInit) {
-    auto arg = cast<DagInit>(intrType);
-    assert(arg->getNumArgs() == 1 && "Only one arg allowed");
-    auto name = arg->getArgNameStr(0);
-    auto num = nameToOrdinal.lookup(name);
-    os << ((first) ? "" : ", ") << num << "->getType()";
-    first = false;
-  }
-  os << "};\n"
-     << " Function *" << callval
-     << " = Intrinsic::getDeclaration(called->getParent(), "
-        "Intrinsic::"
-     << intrName << ", tys);\n";
-  os << "  auto " << FT << " = " << callval << "->getFunctionType();\n";
-  os << "  auto " << cconv << " = orig->getCallingConv();\n";
-}
-
-// Returns whether value generated is a vector value or not.
-bool handle(raw_ostream &os, Record *pattern, Init *resultTree,
-            std::string builder, StringMap<std::string> &nameToOrdinal,
-            bool lookup) {
-  if (DagInit *resultRoot = dyn_cast<DagInit>(resultTree)) {
-    auto opName = resultRoot->getOperator()->getAsString();
-    auto Def = cast<DefInit>(resultRoot->getOperator())->getDef();
-    if (opName == "DiffeRet" || Def->isSubClassOf("DiffeRet")) {
-      os << "dif";
-      return true;
-    } else if (opName == "ConstantFP" || Def->isSubClassOf("ConstantFP")) {
-      if (resultRoot->getNumArgs() != 1)
-        PrintFatalError(pattern->getLoc(), "only single op constant supported");
-
-      auto *argument = resultRoot->getArg(0);
-
-      auto value = dyn_cast<StringInit>(Def->getValueInit("value"));
-      if (!value)
-        PrintFatalError(pattern->getLoc(), Twine("'value' not defined in ") +
-                                               resultTree->getAsString());
-
-      os << "ConstantFP::get(";
-      if (resultRoot->getArgName(0)) {
-        auto name = resultRoot->getArgName(0)->getAsUnquotedString();
-        auto ord = nameToOrdinal.find(name);
-        if (ord == nameToOrdinal.end())
-          PrintFatalError(pattern->getLoc(), Twine("unknown named operand '") +
-                                                 name + "'" +
-                                                 resultTree->getAsString());
-        os << ord->getValue();
-      } else
-        PrintFatalError(pattern->getLoc(),
-                        Twine("unknown named operand in constantfp") +
-                            resultTree->getAsString());
-      os << "->getType(), \"" << value->getValue() << "\")";
-      return false;
-    } else if (opName == "Shadow" || Def->isSubClassOf("Shadow")) {
-      if (resultRoot->getNumArgs() != 1)
-        PrintFatalError(pattern->getLoc(), "only single op constant supported");
-
-      auto *argument = resultRoot->getArg(0);
-      if (lookup)
-        os << "lookup(";
-      os << "gutils->invertPointerM(" << builder << ", ";
-
-      if (resultRoot->getArgName(0)) {
-        auto name = resultRoot->getArgName(0)->getAsUnquotedString();
-        auto ord = nameToOrdinal.find(name);
-        if (ord == nameToOrdinal.end())
-          PrintFatalError(pattern->getLoc(), Twine("unknown named operand '") +
-                                                 name + "'" +
-                                                 resultTree->getAsString());
-        os << ord->getValue();
-
-      } else
-        PrintFatalError(pattern->getLoc(),
-                        Twine("unknown named operand in shadow") +
-                            resultTree->getAsString());
-      os << ")";
-      if (lookup)
-        os << ", " << builder << ")";
-      return true;
-    }
-
-    os << " ({\n";
-    os << "    Value* args[" << resultRoot->getArgs().size() << "];\n";
-
-    SmallVector<bool, 1> vectorValued;
-    bool anyVector = false;
-
-    size_t idx = 0;
-    StringMap<std::string> oldMaps;
-    for (auto zp :
-         llvm::zip(resultRoot->getArgs(), resultRoot->getArgNames())) {
-      os << " args[" << idx << "] = ";
-      idx++;
-      if (isa<UnsetInit>(std::get<0>(zp)) && std::get<1>(zp)) {
-        auto name = std::get<1>(zp)->getAsUnquotedString();
-        auto ord = nameToOrdinal.find(name);
-        if (ord == nameToOrdinal.end())
-          PrintFatalError(pattern->getLoc(), Twine("unknown named operand '") +
-                                                 name + "'" +
-                                                 resultTree->getAsString());
-        if (!StringRef(ord->getValue()).startswith("__tmp_")) {
-          if (lookup)
-            os << "lookup(";
-          os << "gutils->getNewFromOriginal(";
-        }
-        os << ord->getValue();
-        if (!StringRef(ord->getValue()).startswith("__tmp_")) {
-          os << ")";
-          if (lookup)
-            os << ", " << builder << ")";
-        }
-        os << " ;\n";
-        vectorValued.push_back(false);
-        continue;
-      }
-      vectorValued.push_back(
-          handle(os, pattern, std::get<0>(zp), builder, nameToOrdinal, lookup));
-      os << " ;\n";
-      if (std::get<1>(zp)) {
-        auto name = std::get<1>(zp)->getAsUnquotedString();
-        oldMaps.try_emplace(name, nameToOrdinal[name]);
-        nameToOrdinal[name] = "__tmp_" + name;
-        os << " Value* __tmp_" << name << " = args[" << (idx - 1) << "];\n";
-      }
-
-      anyVector |= vectorValued.back();
-    }
-    for (auto &pair : oldMaps) {
-      if (pair.second.size())
-        nameToOrdinal[pair.getKey()] = pair.second;
-      else
-        nameToOrdinal.erase(pair.getKey());
-    }
-
-    bool isCall = opName == "Call" || Def->isSubClassOf("Call");
-    bool isIntr = opName == "Intrinsic" || Def->isSubClassOf("Intrinsic");
-
-    if (isCall) {
-      getFunction(os, "callval", "FT", "cconv", Def->getValueInit("func"));
-    } else if (isIntr) {
-      auto intrName = Def->getValueAsString("name");
-      auto intrTypes = Def->getValueAsListInit("types");
-      getIntrinsic(os, "callval", "FT", "cconv", intrName, intrTypes,
-                   nameToOrdinal);
-    }
-
-    os << " Value *res = nullptr;\n";
-
-    if (anyVector) {
-      os << " if (gutils->getWidth() == 1) { \n";
-    }
-
-    if (isCall || isIntr) {
-      os << " CallInst *cubcall = cast<CallInst>(" << builder
-         << ".CreateCall(FT, callval, ArrayRef<Value*>({";
-    } else {
-      os << "   res = " << builder << ".Create" << opName << "(";
-    }
-    for (size_t i = 0; i < idx; i++) {
-      if (i > 0)
-        os << ", ";
-      os << "args[" << i << "]";
-    }
-    if (isCall || isIntr)
-      os << "})";
-    os << ")";
-    if (isCall || isIntr)
-      os << ")";
-    os << ";\n";
-    if (isCall) {
-      os << " cubcall->setDebugLoc(gutils->getNewFromOriginal(orig->"
-            "getDebugLoc()));\n";
-      os << " cubcall->setCallingConv(cconv);\n";
-      for (auto *attr : *cast<ListInit>(Def->getValueAsListInit("fnattrs"))) {
-        auto attrDef = cast<DefInit>(attr)->getDef();
-        os << "#if LLVM_VERSION_MAJOR >= 14\n"
-           << " cubcall->addAttributeAtIndex(AttributeList::FunctionIndex, "
-           << "Attribute::"
-           << attrDef->getValueInit("name")->getAsUnquotedString() << ");\n";
-        os << "#else\n"
-           << " cubcall->addAttribute(AttributeList::FunctionIndex, "
-           << "Attribute::"
-           << attrDef->getValueInit("name")->getAsUnquotedString() << ");\n";
-        os << "#endif\n";
-      }
-      os << " res = cubcall;\n";
-    } else if (isIntr) {
-      os << " cubcall->setDebugLoc(gutils->getNewFromOriginal(orig->"
-            "getDebugLoc()));\n";
-      os << " cubcall->setCallingConv(cconv);\n";
-      os << " res = cubcall;\n";
-    }
-    if (anyVector) {
-      os << " } else {\n";
-      os << " for(unsigned int idx=0, W=gutils->getWidth(); idx<W; idx++) {\n";
-
-      if (isCall || isIntr) {
-        os << " CallInst *V = cast<CallInst>(" << builder
-           << ".CreateCall(FT, callval, ArrayRef<Value*>({";
-      } else {
-        os << "   Value *V = " << builder << ".Create" << opName << "(";
-      }
-      for (size_t i = 0; i < idx; i++) {
-        if (i > 0)
-          os << ", ";
-        if (vectorValued[i])
-          os << builder << ".CreateExtractValue(args[" << i << "], {idx})";
-        else
-          os << "args[" << i << "]";
-      }
-      if (isCall || isIntr)
-        os << "})";
-      os << ")";
-      if (isCall || isIntr) {
-        os << ")";
-      }
-      os << ";\n";
-
-      if (isCall) {
-        os << "   "
-              "V->setDebugLoc(gutils->getNewFromOriginal(orig->getDebugLoc()));"
-              "\n";
-        os << "   V->setCallingConv(cconv);\n";
-        for (auto *attr : *cast<ListInit>(Def->getValueAsListInit("fnattrs"))) {
-          auto attrDef = cast<DefInit>(attr)->getDef();
-          os << "#if LLVM_VERSION_MAJOR >= 14\n"
-             << "   V->addAttributeAtIndex(AttributeList::FunctionIndex, "
-                "Attribute::"
-             << attrDef->getValueInit("name")->getAsUnquotedString() << ");\n";
-          os << "#else \n"
-             << "   V->addAttribute(AttributeList::FunctionIndex, "
-                "Attribute::"
-             << attrDef->getValueInit("name")->getAsUnquotedString() << ");\n";
-          os << "#endif \n";
-        }
-      }
-      if (isIntr) {
-        os << "   "
-              "V->setDebugLoc(gutils->getNewFromOriginal(orig->getDebugLoc()));"
-              "\n";
-        os << "   V->setCallingConv(cconv);\n";
-      }
-      os << "   if (res == nullptr) res = "
-            "UndefValue::get(ArrayType::get(V->getType(), "
-            "gutils->getWidth()));\n";
-      os << "   res = " << builder << ".CreateInsertValue(res, V, {idx});\n";
-      os << " }\n }\n";
-    }
-    os << " res; })";
-    return anyVector;
-  }
-
-  PrintFatalError(pattern->getLoc(), Twine("unknown dag"));
-}
-
-void emitFullDerivatives(const RecordKeeper &RK,
-                         raw_ostream &os) {
-  emitSourceFileHeader("Rewriters", os);
-  const auto &patterns = RK.getAllDerivedDefinitions("CallPattern");
-  // Ensure unique patterns simply by appending unique suffix.
-  unsigned rewritePatternCount = 0;
-  std::string baseRewriteName = "GeneratedConvert";
-  for (Record *pattern : patterns) {
-    DagInit *tree = pattern->getValueAsDag("PatternToMatch");
-
-    StringMap<std::string> nameToOrdinal;
-    for (int i = 0, e = tree->getNumArgs(); i != e; ++i)
-      nameToOrdinal[tree->getArgNameStr(i)] =
-          "orig->getOperand(" + std::to_string(i) + ")";
-
-    if (tree->getNameStr().str().size())
-      nameToOrdinal[tree->getNameStr().str()] = "orig";
-
-    for (auto arg : tree->getArgs()) {
-      if (isa<DagInit>(arg))
-        PrintFatalError(pattern->getLoc(),
-                        "only single pattern inputs supported");
-    }
-
-    // Emit RewritePattern for Pattern.
-    ListInit *argOps = pattern->getValueAsListInit("ArgDerivatives");
-
-    os << "  if (";
-
-    bool prev = false;
-    for (auto *nameI : *cast<ListInit>(pattern->getValueAsListInit("names"))) {
-      if (prev)
-        os << " ||\n      ";
-      os << "funcName == " << cast<StringInit>(nameI)->getAsString() << "";
-      prev = true;
-    }
-    os << " ){\n";
-    os << "    if (gutils->knownRecomputeHeuristic.find(orig) !=\n";
-    os << "        gutils->knownRecomputeHeuristic.end()) {\n";
-    os << "        if (!gutils->knownRecomputeHeuristic[orig]) {\n";
-    os << "          gutils->cacheForReverse(BuilderZ, newCall,\n";
-    os << "                                  getIndex(orig, "
-          "CacheType::Self));\n";
-    os << "        }\n";
-    os << "    }\n";
-
-    os << "    eraseIfUnused(*orig);\n";
-    os << "    if (gutils->isConstantInstruction(orig))\n";
-    os << "      return;\n";
-
-    os << "    switch (Mode) {\n";
-    os << "      case DerivativeMode::ForwardModeSplit:\n";
-    os << "      case DerivativeMode::ForwardMode:{\n";
-    os << "        IRBuilder<> Builder2(&call);\n";
-    os << "        getForwardBuilder(Builder2);\n";
-    // TODO
-
-    os << "        Value *res = nullptr;\n";
-
-    for (auto argOpEn : llvm::enumerate(*argOps)) {
-      size_t argIdx = argOpEn.index();
-      os << "        if (!gutils->isConstantValue(orig->getArgOperand("
-         << argIdx << "))) {\n";
-      os << "          Value *dif = diffe(orig->getArgOperand(" << argIdx
-         << "), Builder2);\n";
-      DagInit *resultTree = cast<DagInit>(argOpEn.value());
-      os << "          Value *tmp = ";
-
-      bool vectorValued = handle(os, pattern, resultTree, "Builder2",
-                                 nameToOrdinal, /*lookup*/ false);
-      os << ";\n";
-
-      os << "          if (res == nullptr) res = tmp;\n";
-      os << "          else if (gutils->getWidth() == 1) res = "
-            "Builder2.CreateFAdd(res, tmp);\n";
-      os << "          else {\n";
-      if (vectorValued)
-        os << "            Value *out = UndefValue::get(res->getType());\n";
-      else
-        os << "            Value *out = "
-              "UndefValue::get(gutils->getShadowType(res->getType()));\n";
-
-      os << "            for(unsigned int idx=0, W=gutils->getWidth(); idx<W; "
-            "idx++) {\n";
-      os << "              Value *V = "
-            "Builder2.CreateFAdd(Builder2.CreateExtractValue(res, {idx}), ";
-      if (vectorValued)
-        os << "Builder2.CreateExtractValue(tmp, {idx})";
-      else
-        os << "tmp";
-      os << ");\n";
-      os << "              out = Builder2.CreateInsertValue(out, V, {idx});\n";
-      os << "            }\n";
-      os << "            res = out;\n";
-      os << "          }\n";
-      os << "        }\n";
-    }
-
-    os << "        setDiffe(orig, res, Builder2);\n";
-
-    os << "        break;\n";
-    os << "      }\n";
-
-    os << "      case DerivativeMode::ReverseModeGradient:\n";
-    os << "      case DerivativeMode::ReverseModeCombined:{\n";
-    os << "        IRBuilder<> Builder2(&call);\n";
-    os << "        getReverseBuilder(Builder2);\n";
-    // TODO vector
-
-    os << "        Value *dif = nullptr;\n";
-    bool seen = false;
-    for (auto argOpEn : llvm::enumerate(*argOps)) {
-      size_t argIdx = argOpEn.index();
-      os << "        ";
-      if (seen)
-        os << "} else ";
-      seen = true;
-      os << "if (!dif && !gutils->isConstantValue(orig->getArgOperand("
-         << argIdx << "))) {\n";
-      DagInit *resultTree = cast<DagInit>(argOpEn.value());
-      if (hasDiffeRet(resultTree)) {
-        os << "          dif = diffe(orig, Builder2);\n";
-        os << "          setDiffe(orig, "
-              "Constant::getNullValue(gutils->getShadowType(orig->getType())), "
-              "Builder2);\n";
-      }
-    }
-    if (seen)
-      os << "        }\n";
-
-    for (auto argOpEn : llvm::enumerate(*argOps)) {
-      size_t argIdx = argOpEn.index();
-      DagInit *resultTree = cast<DagInit>(argOpEn.value());
-
-      os << "        if (!gutils->isConstantValue(orig->getArgOperand("
-         << argIdx << "))) {\n";
-      os << "          Value *tmp = ";
-      bool vectorValued = handle(os, pattern, resultTree, "Builder2",
-                                 nameToOrdinal, /*lookup*/ true);
-      os << ";\n";
-      os << "          Value *toadd = tmp;\n";
-
-      if (!vectorValued) {
-        os << "          if (gutils->getWidth() > 1) {\n";
-        os << "            toadd = "
-              "UndefValue::get(gutils->getShadowType(tmp->getType()));\n";
-        os << "            for(unsigned int idx=0, W=gutils->getWidth(); "
-              "idx<W; idx++) {\n";
-        os << "              toadd = Builder2.CreateInsertValue(toadd, tmp, "
-              "{idx});\n";
-        os << "            }\n";
-        os << "          }\n";
-      }
-
-      os << "          addToDiffe(orig->getArgOperand(" << argIdx << "), toadd";
-      os << ", Builder2, orig->getArgOperand(" << argIdx << ")->getType());\n";
-      os << "        }\n";
-    }
-
-    os << "        break;\n";
-    os << "      }\n";
-
-    os << "      case DerivativeMode::ReverseModePrimal:{\n";
-    // TODO
-    os << "        break;\n";
-    os << "      }\n";
-    os << "    }\n";
-
-    os << "    return;\n  }\n";
-  }
-}
-
-
-/////////// blas from here on
 
 
 void emitEnumMatcher(const std::vector<Record *> &blas_modes, raw_ostream &os) {
@@ -932,6 +436,7 @@ void emit_helper(Record *pattern, std::vector<size_t> actArgs,
     os 
 << "\n";
   }
+  
 
   os 
 << "  int num_active_fp = 0;\n";
@@ -1022,7 +527,8 @@ llvm::SmallString<40> call_arg_helper(DagInit *argOps,
         llvm::errs() << "MagicInst\n";
       } else if (Def->isSubClassOf("Constant")) {
         auto val = Def->getValueAsString("value");
-        result.append((Twine("ConstantFP::get(innerType, ") + val + ")").str());
+        //result.append((Twine("ConstantFP::get(innerType, ") + val + ")").str());
+        result.append((Twine("ConstantFP::get(fpType, ") + val + ")").str());
       } else {
         llvm::errs() << Def->getName() << "\n";
         PrintFatalError("Def that isn't a DiffeRet!");
@@ -1080,7 +586,8 @@ void emit_deriv_fnc(DagInit *resultTree, const llvm::DenseMap<StringRef, StringR
     else 
       handled.insert(dfnc_name);
     auto retTy = "Builder2.getVoidTy()";
-    if (dfnc_name == "dot") {
+    // TODO: add this to .td file and generate it based on that
+    if (dfnc_name == "dot" || dfnc_name == "asum") {
       retTy = "fpType";
     }
     os
